services/signcalc-service/pyproject.toml
[project]
name = "apex-signcalc-service"
version = "0.1.0"
description = "APEX Sign Calculation Service (wind → member → foundation → anchors → report)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "pydantic>=2.7",
  "orjson>=3.9",
  "uvicorn>=0.30",
  "numpy>=1.26",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "PyYAML>=6.0",
  "ezdxf>=1.2",
  "weasyprint>=61.2",
]

[tool.uv]
dev-dependencies = [
  "pytest>=7.4",
  "httpx>=0.27",
]


